% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Down_uva.R
\name{Down_uva}
\alias{Down_uva}
\title{Download filtering}
\usage{
Down_uva(date_start, date_end, PVC_list, device_id_list, file_s, db.file)
}
\arguments{
\item{date_start}{a character element indicating the date time to start thee download of all the selected devices. The format has to be \%Y-\%m-\%d \%h:\%M:\%S, e.g."2023-11-01 00:00:00"}

\item{date_end}{a character element indicating the date time to end thee download of all the selected devices. The format has to be \%Y-\%m-\%d \%h:\%M:\%S, e.g."2023-11-01 00:00:00"}

\item{PVC_list}{vector containing the "PVC" code of all the devices to download. It should be provided if device_id_list is not.}

\item{device_id_list}{vector containing the device numbre of all the devices to download. It should be provided if PVC_list is not.}

\item{file_s}{path to save the result. If it is not provided, the result is not saved on a csv.}

\item{db.file}{name chosen for your data source when setting up the ODBC connection according to https://wiki.e-ecology.nl/index.php/How_to_access_the_e-Ecology_database}
}
\value{
a dataframe with the downloaded data plus a csv file containing those data if file_s if used.
}
\description{
Download filtering
}
\examples{
DataDownloaded <- Down_uva(date_start = "2023-11-01 00:00:00",date_end = "2023-11-10 00:00:00",PVC_list = c("2X5","3FA"), file_s = "Todaydata.csv", db.file = "GPS_UvA")
}
